Swapped ordering of subsurvey chain in survey references.
I misinterpreted the therion handook, which is not very clear of this.
In a real world test i got wrong results and have concluded the other ordering
based on that data mismatch.
References deeper than one level ("1@b.a") could not be interpreted by
therion compiler.
The reference chain is interpreted right-to-left; with the rightmost survey
must be a child of the current survey context. the dots are just to be
interpreted as an alternative @-sign.