Fixed a bug with structuredWriter wrongly assigning lines after "grade" definition in centerline to central grade.th file.
This bug is a nasty one. In the centerline you can tag the data with a grade-definition referencing a centrally defined
"grade/endgrade" block. However, both commands are named the same but have different functions.
the grade/endgrade stuff should go to a centralized file by default. The grade-reference in the centerline however must
stay there.
Solution/Workaround is, that any grade-command encountered in global context is pushed to the central file, grade commands
in survey/centerline context remain local. This should yield desired behaviour in most cases. It creates a bug whit
grade/endgrade definitions in centerlie context, such definitions remain local to the centerline.